LayoutTests/fast/js/dom-static-property-for-in-iteration-expected.txt

Checks that get_by_pname doesn't get confused about which properties go where when it comes to stacic properties.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'file://' is 'file://'
PASS '' is ''
PASS 'file:///Volumes/Data/WebKit-svn-01/OpenSource/LayoutTests/fast/js/bar' is 'file:///Volumes/Data/WebKit-svn-01/OpenSource/LayoutTests/fast/js/bar'
PASS '' is ''
PASS 'nerget' is 'nerget'
PASS '' is ''
PASS '' is ''
PASS '' is ''
PASS '' is ''
PASS '' is ''
PASS '' is ''
PASS '' is ''
PASS '/Volumes/Data/WebKit-svn-01/OpenSource/LayoutTests/fast/js/bar' is '/Volumes/Data/WebKit-svn-01/OpenSource/LayoutTests/fast/js/bar'
PASS '' is ''
PASS '' is ''
PASS '' is ''
PASS '' is ''
PASS '' is ''
PASS '' is ''
PASS 'file:' is 'file:'
PASS '<a id="foo" href="bar">nerget</a>' is '<a id="foo" href="bar">nerget</a>'
PASS 'true' is 'true'
PASS '' is ''
PASS 'foo' is 'foo'
PASS '' is ''
PASS '' is ''
PASS 'true' is 'true'
PASS 'false' is 'false'
PASS 'nerget' is 'nerget'
PASS '' is ''
PASS 'nerget' is 'nerget'
PASS 'inherit' is 'inherit'
PASS '0' is '0'
PASS 'true' is 'true'
PASS 'nerget' is 'nerget'
PASS '' is ''
PASS '[object HTMLCollection]' is '[object HTMLCollection]'
PASS 'false' is 'false'
PASS '[object CSSStyleDeclaration]' is '[object CSSStyleDeclaration]'
PASS '[object DOMStringMap]' is '[object DOMStringMap]'
PASS '0' is '0'
PASS '0' is '0'
PASS '[object NamedNodeMap]' is '[object NamedNodeMap]'
PASS 'undefined' is 'undefined'
PASS '39' is '39'
PASS '' is ''
PASS '8' is '8'
PASS '' is ''
PASS '0' is '0'
PASS 'null' is 'null'
PASS '[object HTMLBodyElement]' is '[object HTMLBodyElement]'
PASS '[object HTMLScriptElement]' is '[object HTMLScriptElement]'
PASS 'A' is 'A'
PASS '[object HTMLDivElement]' is '[object HTMLDivElement]'
PASS '0' is '0'
PASS '0' is '0'
PASS 'null' is 'null'
PASS '0' is '0'
PASS '18' is '18'
PASS '0' is '0'
PASS '1014' is '1014'
PASS '0' is '0'
PASS '0' is '0'
PASS '[object Text]' is '[object Text]'
PASS '[object HTMLBodyElement]' is '[object HTMLBodyElement]'
PASS '[object Text]' is '[object Text]'
PASS 'file:///Volumes/Data/WebKit-svn-01/OpenSource/LayoutTests/fast/js/dom-static-property-for-in-iteration.html' is 'file:///Volumes/Data/WebKit-svn-01/OpenSource/LayoutTests/fast/js/dom-static-property-for-in-iteration.html'
PASS '[object Text]' is '[object Text]'
PASS 'null' is 'null'
PASS 'nerget' is 'nerget'
PASS '1' is '1'
PASS 'A' is 'A'
PASS 'null' is 'null'
PASS '[object NodeList]' is '[object NodeList]'
PASS '[object Text]' is '[object Text]'
PASS '[object HTMLDocument]' is '[object HTMLDocument]'
PASS 'http://www.w3.org/1999/xhtml' is 'http://www.w3.org/1999/xhtml'
PASS 'a' is 'a'
PASS '[object HTMLBodyElement]' is '[object HTMLBodyElement]'
PASS '1' is '1'
PASS '2' is '2'
PASS '3' is '3'
PASS '4' is '4'
PASS '5' is '5'
PASS '6' is '6'
PASS '1' is '1'
PASS '12' is '12'
PASS '4' is '4'
PASS '1' is '1'
PASS '1' is '1'
PASS '6' is '6'
PASS '3' is '3'
PASS '5' is '5'
PASS '32' is '32'
PASS '11' is '11'
PASS '7' is '7'
PASS '2' is '2'
PASS '10' is '10'
PASS '8' is '8'
PASS '4' is '4'
PASS '2' is '2'
PASS '16' is '16'
PASS '9' is '9'
PASS '8' is '8'
PASS successfullyParsed is true

TEST COMPLETE
nerget
